[CON-43] Clean up more
